fix: cluster DD — PL-V1.38-4: WSL DrvFS detector unification (#1463)
Two divergent WSL-DrvFS detectors lived in:
- `src/config.rs::is_wsl_drvfs_path` — hard-coded `/mnt/<letter>/`,
  used by `coarse_fs_resolution` for mtime-granularity policy.
- `src/cli/watch/mod.rs::is_under_wsl_automount` — read `/etc/wsl.conf`
  for `automount.root` (e.g. `/win/`), used by the watch loop's
  poll-mode detector.

Operators with custom `automount.root=/win/` had the watch loop
correctly trigger `--poll` for `/win/c/...` paths, but
`coarse_fs_resolution` returned 0 (Linux ext4 default) instead of 2s.
Strict mtime-equality skip then silently dropped every rapid second-
save — the bug class PB-V1.30.1-5 / #1225 was supposed to close on
Linux/macOS but slipped through here when only one side learned about
wsl.conf.

Promoted `parse_wsl_automount_root` + a new `wsl_automount_root_or_default`
helper to `crate::config`. Both helpers now share one OnceLock-cached
source of truth.

`is_wsl_drvfs_path` now:
1. First checks the configured automount root (e.g. `/win/<letter>/`).
2. Falls back to the hardcoded `/mnt/<letter>/` if the configured root
   differs from `/mnt/` — operators sometimes mount both paths.
3. Recognises the `//wsl.localhost/` and `//wsl$/` UNC shapes.

`is_under_wsl_automount` collapses to a single delegating call.

Verified: cargo build, cargo test --lib config:: (49/49), cargo test
--bin cqs cli::watch (92/92, 3 ignored), cargo clippy -- -D warnings,
cargo fmt --check.
